pyproject.toml
[project]
name = "azure-nc-llm-demo"
version = "0.1.0"
description = "Demo project for fine-tuning and inference benchmarking on Azure NC H100/H200 GPUs"
authors = ["Azure Demo Team <demo@example.com>"]
requires-python = ">=3.11"
dependencies = [
  "torch>=2.4",
  "transformers>=4.45",
  "accelerate>=0.33",
  "datasets>=3.0",
  "evaluate>=0.4",
  "peft>=0.12",
  "bitsandbytes>=0.43",
  "deepspeed>=0.15",
  "unsloth>=0.4",
  "trl>=0.9",
  "numpy>=1.26",
  "pandas>=2.2",
  "scikit-learn>=1.5",
  "psutil>=5.9",
  "nvidia-ml-py>=12.550",
  "prometheus-client>=0.20",
  "mlflow>=2.14",
  "uvicorn>=0.30",
  "fastapi>=0.112",
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "plotly>=5.24",
  "rich>=13.7",
  "tqdm>=4.66",
  "protobuf>=4.25",
  "tensorboard>=2.16",
  "safetensors>=0.4",
  "sentencepiece>=0.2",
  "huggingface-hub>=0.23",
  "azure-identity>=1.17",
  "azure-ai-ml>=1.16",
  "azure-storage-blob>=12.19",
  "azure-monitor-query>=1.3",
  "vllm>=0.4",
  "sglang>=0.2",
  "lmdeploy>=0.5",
  "tensorrt-llm>=0.10",
  "click>=8.1",
  "typer>=0.12",
  "jinja2>=3.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3",
  "pytest-asyncio>=0.23",
  "ruff>=0.5",
  "mypy>=1.10"
]

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
managed = true

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
known_first_party = ["training", "inference", "metrics", "infra"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "B", "UP", "ASYNC"]
ignore = ["E203"]

